% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.dataset.R
\name{filter.dataset}
\alias{filter.dataset}
\title{Quality control of genotype call files (VCF, BCF, PLINK files)}
\usage{
\method{filter}{dataset}(
  input.file = input,
  plink.files = FALSE,
  bed.file = bed,
  bim.file = bim,
  fam.file = fam,
  remove.related = FALSE,
  kinship.coefficient = NULL,
  geno.value = geno,
  maf.value = maf,
  mind.value = mind,
  limit.LD = FALSE,
  indep.pairwise.kb = NULL,
  indep.pairwise.ct = NULL,
  indep.pairwise.r2 = NULL
)
}
\arguments{
\item{input.file}{(character) The file path to the genetic dataset. Accepts vcf or bcf files.}

\item{plink.files}{(logical) If TRUE, PLINK files will be used (bed, bim, fam).}

\item{bed.file}{(character) The file path for the bed file. Should be provided if plink.files = TRUE}

\item{bim.file}{(character) The file path for the bim file. Should be provided if plink.files = TRUE}

\item{fam.file}{(character) The file path for the fam file. Should be provided if plink.files = TRUE}

\item{remove.related}{(logical) If TRUE, function will calculate the kinship coefficient (to be supplied) of samples and remove those statistically likely to be related. Default is FALSE.}

\item{kinship.coefficient}{(optional) The kinship threshold. Based on kingrelatedness.com/manual, a kinship coefficient range of >0.354 corresponds to duplicate/MZ twins, [0.177, 0.354] to 1st-degree relationships, [0.0844, 0.177] to 2nd-degree relationships, and [0.0442, 0.0884] to 3rd-degree relationships.}

\item{geno.value}{(integer) The percentage threshold of individuals with a missing genotypes to be filtered.}

\item{maf.value}{(integer) The minor allele frequency threshold.}

\item{mind.value}{(integer) The percentage threshold of individuals with missing data to be removed.}

\item{limit.LD}{(logical) If TRUE, function should prune the markers that are in linkage equilibrium. Default is FALSE}

\item{indep.pairwise.kb}{(integer) should be set if limit.LD is TRUE. This is the window size in variant count or kilobase units.}

\item{indep.pairwise.ct}{(integer) should be set if limit.LD is TRUE. This is the variant count to shift the window at each end of step.}

\item{indep.pairwise.r2}{(integer) should be set if limit.LD is TRUE. This is the pairwise threshold. At each step, pairs of variants in the current window with squared correlation greater than the threshold are noted and the variants are pruned from the window until no such pairs remain.}
}
\description{
This function performs quality control of large-scale datasets with multiple filters.
}
\examples{
filter.dataset("test_data/test.csv", plink.files = FALSE, remove.related = TRUE, kinship.coefficient = 0.0844, geno.value = 0.5, maf.value = 0.05, mind.value = 0.1, limit.LD = FALSE)
filter.dataset("test_data/kgpdata.vcf", plink.files = FALSE, remove.related = FALSE, geno.value = 0.5, maf.value = 0.05, mind.value = 0.1, limit.LD = TRUE, indep.pairwise.kb = 200, indep.pairwise.ct = 30, indep.pairwise.r2 = 0.5)
filter.dataset(plink.files = TRUE, bed.file = "mybed.bed", bim.file = "mybim.bim", fam.file = "myfam.fam", remove.related = FALSE, geno.value = 0.5, maf.value = 0.05, mind.value = 0.1, limit.LD = TRUE, indep.pairwise.kb = 300, indep.pairwise.ct = 25, indep.pairwise.r2 = 0.4)
}
